pyproject.toml
[tool.poetry]
name = "motobot_v2"
version = "0.1.0"
description = ""
authors = ["chepurnyi-oo <сhepurnyi_oo@groupbwt.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.37"
pathlib = "^1.0.1"
python-telegram-bot = "^13.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

